+++ src/lib/Timeline.js
@@ -158,6 +158,8 @@ export default class ReactCalendarTimeline extends Component {
       removeListener: PropTypes.func
     }),
 
+    verticalLineClassNamesForTime: PropTypes.func,
+
     children: PropTypes.node
   }
 
@@ -202,6 +204,8 @@ export default class ReactCalendarTimeline extends Component {
     onCanvasMouseLeave: null,
     onCanvasMouseMove: null,
     onZoom: null,
+
+    verticalLineClassNamesForTime: null,
 
     moveResizeValidator: null,
 
@@ -801,6 +805,7 @@ export default class ReactCalendarTimeline extends Component {
         minUnit={minUnit}
         timeSteps={timeSteps}
         height={height}
+        verticalLineClassNamesForTime={this.props.verticalLineClassNamesForTime}
       />
     )
   }